LIBFCREPO-58. Refactoring class path. - Whoopse. - Including a few comment tweaks. - Also restored one additional log entry.
https://issues.umd.edu/browse/LIBFCREPO-58
